Moved worker_manager to managers package.
