+++ aimmo-game/service.py
@@ -15,7 +15,7 @@ from flask_socketio import SocketIO
 from simulation.managers.turn_manager import state_provider, ConcurrentTurnManager
 from simulation import map_generator
 from simulation.avatar.avatar_manager import AvatarManager
-from simulation.worker_manager import WORKER_MANAGERS
+from simulation.managers.worker_manager import WORKER_MANAGERS
 from simulation.world_state import WorldState
 
 app = flask.Flask(__name__)
