+++ aimmo-game/tests/test_simulation/test_worker_manager.py
@@ -7,7 +7,7 @@ from httmock import HTTMock
 
 from simulation.avatar.avatar_manager import AvatarManager
 from simulation.game_state import GameState
-from simulation.worker_manager import WorkerManager
+from simulation.managers.worker_manager import WorkerManager
 from .maps import InfiniteMap
